refactor(autonomi): get rid of the native api
