--- autonomi/src/native/mod.rs
@@ -1,5 +1,0 @@
-pub use crate::client::Client;
-
-pub mod client;
-#[cfg(feature = "transfers")]
-pub mod wallet;
