--- autonomi/tests/evm/mod.rs
@@ -1,7 +1,0 @@
-#[cfg(feature = "files")]
-mod file;
-#[cfg(feature = "data")]
-mod put;
-#[cfg(feature = "registers")]
-mod register;
-mod wallet;
