--- autonomi/tests/integration.rs
@@ -1,5 +1,0 @@
-pub(crate) mod common;
-#[cfg(feature = "evm-payments")]
-mod evm;
-#[cfg(feature = "native-payments")]
-mod native;
