--- autonomi/tests/native/mod.rs
@@ -1,10 +1,0 @@
-use autonomi;
-
-#[cfg(feature = "files")]
-mod file;
-#[cfg(feature = "data")]
-mod put;
-#[cfg(feature = "registers")]
-mod register;
-
-pub type Client = autonomi::native::client::NativeClient;
